+++ standard-deviation.go
@@ -6,7 +6,7 @@ import (
 
 // This function calculates the standart deviation
 // The SD is the square root of the variance
-func Standard_Deviation(variance float64) int {
+func Standard_Deviation(variance float64) float64 {
 	deviation := math.Sqrt(variance)
-	return int(math.Round(deviation))
+	return float64(int(math.Round(deviation)))
 }
